+++ frontend/src/components/common/layout/Navbar.tsx
@@ -3,8 +3,6 @@
 import { useRef, useState } from "react";
 
 //Next
-import Image from "next/image";
-import Link from "next/link";
 import { useSession } from "next-auth/react";
 
 //Hooks
@@ -16,11 +14,7 @@ import { useSidebarsContext } from "@/context/SidebarsContext";
 //Components
 import ToggleTheme from "../ui/buttons/ToggleTheme";
 //UI
-import { FaGithub, FaLinkedin } from "react-icons/fa";
 import { Button } from "@/components/ui/button";
-import logo from "/public/logo.png";
-import { contributors } from "@/constants";
-import UserAvatar from "@/components/common/ui/image/UserAvatar";
 import NavbarLinks from "./NavbarLinks";
 import { Menu, SquareUserRound } from "lucide-react";
 import NavbarAsozialMenu from "./NavbarAsozialMenu";
