+++ frontend/src/components/common/layout/NavbarAsozialMenu.tsx
@@ -7,7 +7,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import Link from "next/link";
-import AsozialLogo from "../ui/AsozialLogo";
+import AsozialLogo from "../ui/image/AsozialLogo";
 import ContributorsImages from "./ContributorsImages";
 import { Button } from "@/components/ui/button";
