+++ frontend/src/components/project/ProjectSidebar.tsx
@@ -8,7 +8,7 @@ import { useSession } from "next-auth/react";
 import SidebarTitle from "../common/sidebar/SidebarTitle";
 import SearchIcon from "../common/ui/icons/SearchIcon";
 import MatchIcon from "../common/ui/icons/MatchIcon";
-import ProjectCreateIcon from "../common/ui/ProjectCreateIcon";
+import ProjectCreateIcon from "../common/ui/icons/ProjectCreateIcon";
 import SidebarButtons from "../common/sidebar/SidebarButtons";
 import SidebarFriendsList from "../common/sidebar/SidebarFriendsList";
